issue #70 separate evaluation from scoring (#80)
* issue #70 separate evaluation from scoring
