+++ README.md
@@ -1,4 +1,4 @@
-[![Build Status](https://dev.azure.com/teradata-consulting/AnalyticOps/_apis/build/status/ThinkBigAnalytics.AoaDemoModels?branchName=master)](https://dev.azure.com/teradata-consulting/AnalyticOps/_build/latest?definitionId=95&branchName=master)
+
 
 # Models
