+++ byom/pima/requirements.txt
@@ -1,0 +1,3 @@
+scikit-learn>=0.22.2
+matplotlib>=2.2.2
+teradataml>=17.0.0.0
