+++ model_definitions/03c9a01f-bd46-4e7c-9a60-4282039094e6/model_modules/evaluation.py
@@ -1,0 +1,63 @@
+from sklearn import metrics
+from teradataml import create_context
+from teradataml.dataframe.dataframe import DataFrame
+
+import os
+import joblib
+import json
+
+
+def save_plot(title):
+    import matplotlib.pyplot as plt
+
+    plt.title(title)
+    fig = plt.gcf()
+    filename = title.replace(" ", "_").lower()
+    fig.savefig('artifacts/output/{}'.format(filename), dpi=500)
+    plt.clf()
+
+
+def evaluate(data_conf, model_conf, **kwargs):
+    model = joblib.load('artifacts/input/model.joblib')
+
+    create_context(host=os.environ["AOA_CONN_HOST"],
+                   username=os.environ["AOA_CONN_USERNAME"],
+                   password=os.environ["AOA_CONN_PASSWORD"])
+
+    # Read test dataset from Teradata
+    # As this is for demo purposes, we simulate the test dataset changing between executions
+    # by introducing a random sample. Note that the sampling is performed in Teradata!
+    test_df = DataFrame(data_conf["table"]).sample(frac=0.8)
+    test_df = test_df.to_pandas()
+
+    X_test = test_df[model.feature_names]
+    y_test = test_df[model.target_name]
+
+    print("Scoring")
+    y_pred = model.predict(test_df[model.feature_names])
+
+    evaluation = {
+        'Accuracy': '{:.2f}'.format(metrics.accuracy_score(y_test, y_pred)),
+        'Recall': '{:.2f}'.format(metrics.recall_score(y_test, y_pred)),
+        'Precision': '{:.2f}'.format(metrics.precision_score(y_test, y_pred)),
+        'f1-score': '{:.2f}'.format(metrics.f1_score(y_test, y_pred))
+    }
+
+    with open("artifacts/output/metrics.json", "w+") as f:
+        json.dump(evaluation, f)
+
+    metrics.plot_confusion_matrix(model, X_test, y_test)
+    save_plot('Confusion Matrix')
+
+    metrics.plot_roc_curve(model, X_test, y_test)
+    save_plot('ROC Curve')
+
+    # xgboost has its own feature importance plot support but lets use shap as explainability example
+    import shap
+
+    shap_explainer = shap.TreeExplainer(model['xgb'])
+    shap_values = shap_explainer.shap_values(X_test)
+
+    shap.summary_plot(shap_values, X_test, feature_names=model.feature_names,
+                      show=False, plot_size=(12,8), plot_type='bar')
+    save_plot('SHAP Feature Importance')
