+++ model_definitions/149e31ed-c554-46b4-95d2-00c5c43320fb/model_modules/evaluation.py
@@ -1,0 +1,66 @@
+from pyspark import SparkConf
+from pyspark.sql import SparkSession
+from sklearn import metrics
+from .util import read_dataframe
+
+import logging
+import joblib
+import json
+
+logging.getLogger("py4j").setLevel(logging.ERROR)
+
+# spark name and other properties are set by the framework launcher in spark-submit. don't override
+spark = SparkSession.builder \
+    .config(conf=SparkConf()) \
+    .getOrCreate()
+
+
+def save_plot(title):
+    import matplotlib.pyplot as plt
+
+    plt.title(title)
+    fig = plt.gcf()
+    filename = title.replace(" ", "_").lower()
+    fig.savefig('artifacts/output/{}'.format(filename), dpi=500)
+    plt.clf()
+
+
+def evaluate(data_conf, model_conf, **kwargs):
+    model = joblib.load('artifacts/input/model.joblib')
+
+    test_df = read_dataframe(spark, data_conf["url"])
+
+    # do feature eng in spark / joins whatever reason you're using pyspark...
+    # split into test and train
+    test_df = test_df.randomSplit([0.7, 0.3], 42)[1].toPandas()
+
+    X_test = test_df[model.feature_names]
+    y_test = test_df["HasDiabetes"]
+
+    y_pred = model.predict(X_test)
+
+    evaluation = {
+        'Accuracy': '{:.2f}'.format(metrics.accuracy_score(y_test, y_pred)),
+        'Recall': '{:.2f}'.format(metrics.recall_score(y_test, y_pred)),
+        'Precision': '{:.2f}'.format(metrics.precision_score(y_test, y_pred)),
+        'f1-score': '{:.2f}'.format(metrics.f1_score(y_test, y_pred))
+    }
+
+    with open("artifacts/output/metrics.json", "w+") as f:
+        json.dump(evaluation, f)
+
+    metrics.plot_confusion_matrix(model, X_test, y_test)
+    save_plot('Confusion Matrix')
+
+    metrics.plot_roc_curve(model, X_test, y_test)
+    save_plot('ROC Curve')
+
+    # xgboost has its own feature importance plot support but lets use shap as explainability example
+    import shap
+
+    shap_explainer = shap.TreeExplainer(model['xgb'])
+    shap_values = shap_explainer.shap_values(X_test)
+
+    shap.summary_plot(shap_values, X_test, feature_names=model.feature_names,
+                      show=False, plot_size=(12,8), plot_type='bar')
+    save_plot('SHAP Feature Importance')
