+++ model_definitions/149e31ed-c554-46b4-95d2-00c5c43320fb/model_modules/scoring.py
@@ -1,12 +1,10 @@
 from pyspark import SparkConf
 from pyspark.sql import SparkSession
-from sklearn import metrics
+from .util import read_dataframe
 
 import logging
 import joblib
-import urllib.request
 import pandas as pd
-import json
 import os
 
 
@@ -18,22 +16,10 @@ spark = SparkSession.builder \
     .getOrCreate()
 
 
-def read_dataframe(url):
-    # in a real world scenario, you would read from S3, HDFS, Teradata,
-    # etc but for demo reading from url. we could read via pandas.read_csv but just to show pyspark ...
-    urllib.request.urlretrieve(url, "/tmp/data.csv")
-    column_names = ["NumTimesPrg", "PlGlcConc", "BloodP", "SkinThick", "TwoHourSerIns", "BMI", "DiPedFunc", "Age", "HasDiabetes"]
-
-    return spark.read.format("csv")\
-        .option("inferSchema", "true")\
-        .load("/tmp/data.csv")\
-        .toDF(*column_names)
-
-
 def score(data_conf, model_conf, **kwargs):
     model = joblib.load('artifacts/input/model.joblib')
 
-    test_df = read_dataframe(data_conf["url"])
+    test_df = read_dataframe(spark, data_conf["url"])
 
     # do feature eng in spark / joins whatever reason you're using pyspark...
     # split into test and train
@@ -53,57 +39,6 @@ def score(data_conf, model_conf, **kwargs):
     predictions.write.mode("overwrite").save("/tmp/predictions")
 
     logging.info("Finished scoring")
-
-
-def save_plot(title):
-    import matplotlib.pyplot as plt
-
-    plt.title(title)
-    fig = plt.gcf()
-    filename = title.replace(" ", "_").lower()
-    fig.savefig('artifacts/output/{}'.format(filename), dpi=500)
-    plt.clf()
-
-
-def evaluate(data_conf, model_conf, **kwargs):
-    model = joblib.load('artifacts/input/model.joblib')
-
-    test_df = read_dataframe(data_conf["url"])
-
-    # do feature eng in spark / joins whatever reason you're using pyspark...
-    # split into test and train
-    test_df = test_df.randomSplit([0.7, 0.3], 42)[1].toPandas()
-
-    X_test = test_df[model.feature_names]
-    y_test = test_df["HasDiabetes"]
-
-    y_pred = model.predict(X_test)
-
-    evaluation = {
-        'Accuracy': '{:.2f}'.format(metrics.accuracy_score(y_test, y_pred)),
-        'Recall': '{:.2f}'.format(metrics.recall_score(y_test, y_pred)),
-        'Precision': '{:.2f}'.format(metrics.precision_score(y_test, y_pred)),
-        'f1-score': '{:.2f}'.format(metrics.f1_score(y_test, y_pred))
-    }
-
-    with open("artifacts/output/metrics.json", "w+") as f:
-        json.dump(evaluation, f)
-
-    metrics.plot_confusion_matrix(model, X_test, y_test)
-    save_plot('Confusion Matrix')
-
-    metrics.plot_roc_curve(model, X_test, y_test)
-    save_plot('ROC Curve')
-
-    # xgboost has its own feature importance plot support but lets use shap as explainability example
-    import shap
-
-    shap_explainer = shap.TreeExplainer(model['xgb'])
-    shap_values = shap_explainer.shap_values(X_test)
-
-    shap.summary_plot(shap_values, X_test, feature_names=model.feature_names,
-                      show=False, plot_size=(12,8), plot_type='bar')
-    save_plot('SHAP Feature Importance')
 
 
 # Add code required for RESTful API
